requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
pydantic
